Forked the XOR.
The 6-transistor XOR gate works fine, but depends on having access to the
body terminal of the transistor. In an IC, all body terminals are connected to
VCC (for PMOS) or GND (for NMOS), so that the gate voltages are relative to the
power supply rails. Therefore, you can use NMOS as a high-side switch and PMOS
as a low-side switch.

The 6T gate depends on this, because the power rails for the middle inverter are
not VCC and GND, but A and #A. When A is asserted, the middle inverter rails are
reversed from normal, and turning on the PMOS lets through low voltage instead
of the normal high voltage (and vice versa for the NMOS). You can turn on the
PMOS because the gate referenced to the body, which is tied to VCC, not to
source (and vice versa for the NMOS).

With discrete transistors, the body is always tied to the source. This means
that if you reverse the rails, you can't turn on either of the transistors,
because for the PMOS, you need the gate to be lower than the source, but the
source is low (and vice versa for the NMOS).

It is easy enough to simulate a transistor with its body wired properly -- just
ignore it. On a PMOS, the drain is set to equal the source if the gate is low,
and set to Z if the gate is high. On an NMOS, the drain is set equal to the
source if the gate is high, and Z if low.

However, a circuit built with discrete FETs won't work. As soon as the source
of the PMOS goes low, you have to hit the gate with a negative voltage (lower
than low), and as soon as the source of the NMOS goes high, you have to hit it
with a voltage higher than high.

The CD4007 chip is a set of three complementary pairs of transistors. The first
pair has its source and body pins tied to VCC and GND. The gates are connected
together and the drains are independent. The second pair has its sources
connected to pins, but its body tied to VCC and GND, respectively. The gates are
again connected together and the drains are again independent. The third pair
is specifically wired as an inverter, with its sources connected to pins, gates
connected together, and drains also connected together.

So, for simulation, we use the DiscreteXOR 6T gate, but if we actually want to
implement it, we need either a gate with more transistors, or transistors with
the body terminal tied to VCC/GND, and not necessarily source.